Fixed Context and Thread leaks in App
The reference to the Context passed in App's constructor was never
removed from the static map of Contexts and Apps, resulting
in a lot of leaked activites. Instead of mapping a Context to
null in shutdown() it's now actually removed.
The HandlerThread in App was never terminated, resulting in a lot
of leaked threads. The HandlerThread is now gracefully terminated
in shutdown().

Change-Id: I3fadf594f89acec2a7c87ec39ed30e570beb9dd8